Split networking out from node initialization into its own package
Node and networking are now completely independent initialization paths.
