+++ pkg/cmd/server/kubernetes/node/node.go
@@ -13,34 +13,13 @@ import (
 	dockerclient "github.com/fsouza/go-dockerclient"
 	"github.com/golang/glog"
 
-	"github.com/openshift/origin/pkg/proxy/hybrid"
-	"github.com/openshift/origin/pkg/proxy/unidler"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	utilnet "k8s.io/apimachinery/pkg/util/net"
-	utilwait "k8s.io/apimachinery/pkg/util/wait"
-	kv1core "k8s.io/client-go/kubernetes/typed/core/v1"
-	kclientv1 "k8s.io/client-go/pkg/api/v1"
-	"k8s.io/client-go/tools/record"
 	kubeletapp "k8s.io/kubernetes/cmd/kubelet/app"
-	kapi "k8s.io/kubernetes/pkg/api"
 	kapiv1 "k8s.io/kubernetes/pkg/api/v1"
-	"k8s.io/kubernetes/pkg/apis/componentconfig"
-	kclientset "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
 	"k8s.io/kubernetes/pkg/kubelet/cadvisor"
 	cadvisortesting "k8s.io/kubernetes/pkg/kubelet/cadvisor/testing"
 	"k8s.io/kubernetes/pkg/kubelet/cm"
 	dockertools "k8s.io/kubernetes/pkg/kubelet/dockershim/libdocker"
-	proxy "k8s.io/kubernetes/pkg/proxy"
-	pconfig "k8s.io/kubernetes/pkg/proxy/config"
-	"k8s.io/kubernetes/pkg/proxy/healthcheck"
-	"k8s.io/kubernetes/pkg/proxy/iptables"
-	"k8s.io/kubernetes/pkg/proxy/userspace"
-	utildbus "k8s.io/kubernetes/pkg/util/dbus"
-	kexec "k8s.io/kubernetes/pkg/util/exec"
-	utilexec "k8s.io/kubernetes/pkg/util/exec"
-	utiliptables "k8s.io/kubernetes/pkg/util/iptables"
-	utilnode "k8s.io/kubernetes/pkg/util/node"
-	utilsysctl "k8s.io/kubernetes/pkg/util/sysctl"
 	"k8s.io/kubernetes/pkg/volume"
 
 	configapi "github.com/openshift/origin/pkg/cmd/server/api"
@@ -244,7 +223,7 @@ func (c *NodeConfig) EnsureLocalQuota(nodeConfig configapi.NodeConfig) {
 func (c *NodeConfig) RunKubelet() {
 	var clusterDNS net.IP
 	if len(c.KubeletServer.ClusterDNS) == 0 {
-		if service, err := c.Client.Core().Services(metav1.NamespaceDefault).Get("kubernetes", metav1.GetOptions{}); err == nil {
+		if service, err := c.KubeletDeps.KubeClient.Core().Services(metav1.NamespaceDefault).Get("kubernetes", metav1.GetOptions{}); err == nil {
 			if includesServicePort(service.Spec.Ports, 53, "dns") {
 				// Use master service if service includes "dns" port 53.
 				clusterDNS = net.ParseIP(service.Spec.ClusterIP)
@@ -252,7 +231,7 @@ func (c *NodeConfig) RunKubelet() {
 		}
 	}
 	if clusterDNS == nil {
-		if endpoint, err := c.Client.Core().Endpoints(metav1.NamespaceDefault).Get("kubernetes", metav1.GetOptions{}); err == nil {
+		if endpoint, err := c.KubeletDeps.KubeClient.Core().Endpoints(metav1.NamespaceDefault).Get("kubernetes", metav1.GetOptions{}); err == nil {
 			if endpointIP, ok := firstEndpointIPWithNamedPort(endpoint, 53, "dns"); ok {
 				// Use first endpoint if endpoint includes "dns" port 53.
 				clusterDNS = net.ParseIP(endpointIP)
@@ -301,209 +280,8 @@ func SetFakeContainerManagerInterfaceForIntegrationTest() {
 	defaultContainerManagerInterface = cm.NewStubContainerManager()
 }
 
-// RunSDN starts the SDN, if the OpenShift SDN network plugin is enabled in configuration.
-func (c *NodeConfig) RunSDN() {
-	if c.SDNNode == nil {
-		return
-	}
-	if err := c.SDNNode.Start(); err != nil {
-		glog.Fatalf("error: SDN node startup failed: %v", err)
-	}
-}
-
-// RunDNS starts the DNS server as soon as services are loaded.
-func (c *NodeConfig) RunDNS() {
-	go func() {
-		glog.Infof("Starting DNS on %s", c.DNSServer.Config.DnsAddr)
-		err := c.DNSServer.ListenAndServe()
-		glog.Fatalf("DNS server failed to start: %v", err)
-	}()
-}
-
-// RunProxy starts the proxy
-func (c *NodeConfig) RunProxy() {
-	protocol := utiliptables.ProtocolIpv4
-	bindAddr := net.ParseIP(c.ProxyConfig.BindAddress)
-	if bindAddr.To4() == nil {
-		protocol = utiliptables.ProtocolIpv6
-	}
-
-	portRange := utilnet.ParsePortRangeOrDie(c.ProxyConfig.PortRange)
-
-	hostname := utilnode.GetHostname(c.KubeletServer.HostnameOverride)
-
-	eventBroadcaster := record.NewBroadcaster()
-	eventBroadcaster.StartRecordingToSink(&kv1core.EventSinkImpl{Interface: kv1core.New(c.ExternalKubeClientset.CoreV1().RESTClient()).Events("")})
-	recorder := eventBroadcaster.NewRecorder(kapi.Scheme, kclientv1.EventSource{Component: "kube-proxy", Host: hostname})
-
-	execer := kexec.New()
-	dbus := utildbus.New()
-	iptInterface := utiliptables.New(execer, dbus, protocol)
-
-	var proxier proxy.ProxyProvider
-	var servicesHandler pconfig.ServiceHandler
-	var endpointsHandler pconfig.EndpointsHandler
-
-	switch c.ProxyConfig.Mode {
-	case componentconfig.ProxyModeIPTables:
-		glog.V(0).Info("Using iptables Proxier.")
-		if bindAddr.Equal(net.IPv4zero) {
-			bindAddr = getNodeIP(c.Client, hostname)
-		}
-		var healthzServer *healthcheck.HealthzServer
-		if len(c.ProxyConfig.HealthzBindAddress) > 0 {
-			healthzServer = healthcheck.NewDefaultHealthzServer(c.ProxyConfig.HealthzBindAddress, 2*c.ProxyConfig.IPTables.SyncPeriod.Duration)
-		}
-		if c.ProxyConfig.IPTables.MasqueradeBit == nil {
-			// IPTablesMasqueradeBit must be specified or defaulted.
-			glog.Fatalf("Unable to read IPTablesMasqueradeBit from config")
-		}
-		proxierIptables, err := iptables.NewProxier(
-			iptInterface,
-			utilsysctl.New(),
-			execer,
-			c.ProxyConfig.IPTables.SyncPeriod.Duration,
-			c.ProxyConfig.IPTables.MinSyncPeriod.Duration,
-			c.ProxyConfig.IPTables.MasqueradeAll,
-			int(*c.ProxyConfig.IPTables.MasqueradeBit),
-			c.ProxyConfig.ClusterCIDR,
-			hostname,
-			bindAddr,
-			recorder,
-			healthzServer,
-		)
-
-		if err != nil {
-			if c.Containerized {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy: %v\n When running in a container, you must run the container in the host network namespace with --net=host and with --privileged", err)
-			} else {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy. You must run this process as root to use the service proxy: %v", err)
-			}
-		}
-		proxier = proxierIptables
-		endpointsHandler = proxierIptables
-		servicesHandler = proxierIptables
-		// No turning back. Remove artifacts that might still exist from the userspace Proxier.
-		glog.V(0).Info("Tearing down userspace rules.")
-		userspace.CleanupLeftovers(iptInterface)
-	case componentconfig.ProxyModeUserspace:
-		glog.V(0).Info("Using userspace Proxier.")
-		// This is a proxy.LoadBalancer which NewProxier needs but has methods we don't need for
-		// our config.EndpointsHandler.
-		loadBalancer := userspace.NewLoadBalancerRR()
-		// set EndpointsHandler to our loadBalancer
-		endpointsHandler = loadBalancer
-
-		execer := utilexec.New()
-		proxierUserspace, err := userspace.NewProxier(
-			loadBalancer,
-			bindAddr,
-			iptInterface,
-			execer,
-			*portRange,
-			c.ProxyConfig.IPTables.SyncPeriod.Duration,
-			c.ProxyConfig.IPTables.MinSyncPeriod.Duration,
-			c.ProxyConfig.UDPIdleTimeout.Duration,
-		)
-		if err != nil {
-			if c.Containerized {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy: %v\n When running in a container, you must run the container in the host network namespace with --net=host and with --privileged", err)
-			} else {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy. You must run this process as root to use the service proxy: %v", err)
-			}
-		}
-		proxier = proxierUserspace
-		servicesHandler = proxierUserspace
-		// Remove artifacts from the pure-iptables Proxier.
-		glog.V(0).Info("Tearing down pure-iptables proxy rules.")
-		iptables.CleanupLeftovers(iptInterface)
-	default:
-		glog.Fatalf("Unknown proxy mode %q", c.ProxyConfig.Mode)
-	}
-
-	// Create configs (i.e. Watches for Services and Endpoints)
-	// Note: RegisterHandler() calls need to happen before creation of Sources because sources
-	// only notify on changes, and the initial update (on process start) may be lost if no handlers
-	// are registered yet.
-	serviceConfig := pconfig.NewServiceConfig(
-		c.InternalKubeInformers.Core().InternalVersion().Services(),
-		c.ProxyConfig.ConfigSyncPeriod.Duration,
-	)
-
-	if c.EnableUnidling {
-		unidlingLoadBalancer := userspace.NewLoadBalancerRR()
-		signaler := unidler.NewEventSignaler(recorder)
-		unidlingUserspaceProxy, err := unidler.NewUnidlerProxier(unidlingLoadBalancer, bindAddr, iptInterface, execer, *portRange, c.ProxyConfig.IPTables.SyncPeriod.Duration, c.ProxyConfig.IPTables.MinSyncPeriod.Duration, c.ProxyConfig.UDPIdleTimeout.Duration, signaler)
-		if err != nil {
-			if c.Containerized {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy: %v\n When running in a container, you must run the container in the host network namespace with --net=host and with --privileged", err)
-			} else {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy. You must run this process as root to use the service proxy: %v", err)
-			}
-		}
-		hybridProxier, err := hybrid.NewHybridProxier(
-			unidlingLoadBalancer,
-			unidlingUserspaceProxy,
-			endpointsHandler,
-			servicesHandler,
-			proxier,
-			unidlingUserspaceProxy,
-			c.ProxyConfig.IPTables.SyncPeriod.Duration,
-			c.InternalKubeInformers.Core().InternalVersion().Services().Lister(),
-		)
-		if err != nil {
-			if c.Containerized {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy: %v\n When running in a container, you must run the container in the host network namespace with --net=host and with --privileged", err)
-			} else {
-				glog.Fatalf("error: Could not initialize Kubernetes Proxy. You must run this process as root to use the service proxy: %v", err)
-			}
-		}
-		endpointsHandler = hybridProxier
-		servicesHandler = hybridProxier
-		proxier = hybridProxier
-	}
-
-	iptInterface.AddReloadFunc(proxier.Sync)
-	serviceConfig.RegisterEventHandler(servicesHandler)
-	go serviceConfig.Run(utilwait.NeverStop)
-
-	endpointsConfig := pconfig.NewEndpointsConfig(
-		c.InternalKubeInformers.Core().InternalVersion().Endpoints(),
-		c.ProxyConfig.ConfigSyncPeriod.Duration,
-	)
-	// customized handling registration that inserts a filter if needed
-	if c.SDNProxy != nil {
-		if err := c.SDNProxy.Start(endpointsHandler); err != nil {
-			glog.Fatalf("error: node proxy plugin startup failed: %v", err)
-		}
-		endpointsHandler = c.SDNProxy
-	}
-	endpointsConfig.RegisterEventHandler(endpointsHandler)
-	go endpointsConfig.Run(utilwait.NeverStop)
-
-	// periodically sync k8s iptables rules
-	go utilwait.Forever(proxier.SyncLoop, 0)
-	glog.Infof("Started Kubernetes Proxy on %s", c.ProxyConfig.BindAddress)
-}
-
-// getNodeIP is copied from the upstream proxy config to retrieve the IP of a node.
-func getNodeIP(client kclientset.Interface, hostname string) net.IP {
-	var nodeIP net.IP
-	node, err := client.Core().Nodes().Get(hostname, metav1.GetOptions{})
-	if err != nil {
-		glog.Warningf("Failed to retrieve node info: %v", err)
-		return nil
-	}
-	nodeIP, err = utilnode.InternalGetNodeHostIP(node)
-	if err != nil {
-		glog.Warningf("Failed to retrieve node IP: %v", err)
-		return nil
-	}
-	return nodeIP
-}
-
 // TODO: more generic location
-func includesServicePort(ports []kapi.ServicePort, port int, portName string) bool {
+func includesServicePort(ports []kapiv1.ServicePort, port int, portName string) bool {
 	for _, p := range ports {
 		if p.Port == int32(port) && p.Name == portName {
 			return true
@@ -513,7 +291,7 @@ func includesServicePort(ports []kapi.ServicePort, port int, portName string) bo
 }
 
 // TODO: more generic location
-func includesEndpointPort(ports []kapi.EndpointPort, port int) bool {
+func includesEndpointPort(ports []kapiv1.EndpointPort, port int) bool {
 	for _, p := range ports {
 		if p.Port == int32(port) {
 			return true
@@ -523,7 +301,7 @@ func includesEndpointPort(ports []kapi.EndpointPort, port int) bool {
 }
 
 // TODO: more generic location
-func firstEndpointIP(endpoints *kapi.Endpoints, port int) (string, bool) {
+func firstEndpointIP(endpoints *kapiv1.Endpoints, port int) (string, bool) {
 	for _, s := range endpoints.Subsets {
 		if !includesEndpointPort(s.Ports, port) {
 			continue
@@ -536,7 +314,7 @@ func firstEndpointIP(endpoints *kapi.Endpoints, port int) (string, bool) {
 }
 
 // TODO: more generic location
-func firstEndpointIPWithNamedPort(endpoints *kapi.Endpoints, port int, portName string) (string, bool) {
+func firstEndpointIPWithNamedPort(endpoints *kapiv1.Endpoints, port int, portName string) (string, bool) {
 	for _, s := range endpoints.Subsets {
 		if !includesNamedEndpointPort(s.Ports, port, portName) {
 			continue
@@ -549,7 +327,7 @@ func firstEndpointIPWithNamedPort(endpoints *kapi.Endpoints, port int, portName 
 }
 
 // TODO: more generic location
-func includesNamedEndpointPort(ports []kapi.EndpointPort, port int, portName string) bool {
+func includesNamedEndpointPort(ports []kapiv1.EndpointPort, port int, portName string) bool {
 	for _, p := range ports {
 		if p.Port == int32(port) && p.Name == portName {
 			return true
