+++ pkg/cmd/server/kubernetes/node/node_config.go
@@ -2,39 +2,24 @@ package node
 
 import (
 	"crypto/tls"
-	"fmt"
-	"net"
-	"strings"
 
 	"github.com/golang/glog"
 
-	miekgdns "github.com/miekg/dns"
-
-	kerrs "k8s.io/apimachinery/pkg/api/errors"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 	clientgoclientset "k8s.io/client-go/kubernetes"
 	kv1core "k8s.io/client-go/kubernetes/typed/core/v1"
 	"k8s.io/client-go/util/cert"
 	kubeletapp "k8s.io/kubernetes/cmd/kubelet/app"
 	kubeletoptions "k8s.io/kubernetes/cmd/kubelet/app/options"
-	"k8s.io/kubernetes/pkg/apis/componentconfig"
 	"k8s.io/kubernetes/pkg/apis/componentconfig/v1alpha1"
-	kclientsetexternal "k8s.io/kubernetes/pkg/client/clientset_generated/clientset"
-	kclientset "k8s.io/kubernetes/pkg/client/clientset_generated/internalclientset"
-	kinternalinformers "k8s.io/kubernetes/pkg/client/informers/informers_generated/internalversion"
 	"k8s.io/kubernetes/pkg/cloudprovider"
 	"k8s.io/kubernetes/pkg/kubelet"
 	dockertools "k8s.io/kubernetes/pkg/kubelet/dockershim/libdocker"
 	kubeletserver "k8s.io/kubernetes/pkg/kubelet/server"
 
-	osclient "github.com/openshift/origin/pkg/client"
 	configapi "github.com/openshift/origin/pkg/cmd/server/api"
 	"github.com/openshift/origin/pkg/cmd/server/crypto"
 	cmdutil "github.com/openshift/origin/pkg/cmd/util"
-	"github.com/openshift/origin/pkg/dns"
-	"github.com/openshift/origin/pkg/network"
-	networkapi "github.com/openshift/origin/pkg/network/apis/network"
 )
 
 // NodeConfig represents the required parameters to start the OpenShift node
@@ -49,35 +34,15 @@ type NodeConfig struct {
 	// Containerized is true if we are expected to be running inside of a container
 	Containerized bool
 
-	// Client to connect to the master.
-	Client kclientset.Interface
-	// External kube client
-	ExternalKubeClientset kclientsetexternal.Interface
-	// Internal kubernetes shared informer factory.
-	InternalKubeInformers kinternalinformers.SharedInformerFactory
 	// DockerClient is a client to connect to Docker
 	DockerClient dockertools.Interface
 	// KubeletServer contains the KubeletServer configuration
 	KubeletServer *kubeletoptions.KubeletServer
 	// KubeletDeps are the injected code dependencies for the kubelet, fully initialized
 	KubeletDeps *kubelet.KubeletDeps
-	// ProxyConfig is the configuration for the kube-proxy, fully initialized
-	ProxyConfig *componentconfig.KubeProxyConfiguration
-	// IPTablesSyncPeriod is how often iptable rules are refreshed
-	IPTablesSyncPeriod string
-	// EnableUnidling indicates whether or not the unidling hybrid proxy should be used
-	EnableUnidling bool
-
-	// DNSConfig controls the DNS configuration.
-	DNSServer *dns.Server
-
-	// SDNNode is an optional SDN node interface
-	SDNNode network.NodeInterface
-	// SDNProxy is an optional service endpoints filterer
-	SDNProxy network.ProxyInterface
 }
 
-func New(options configapi.NodeConfig, server *kubeletoptions.KubeletServer, proxyconfig *componentconfig.KubeProxyConfiguration, enableProxy, enableDNS bool) (*NodeConfig, error) {
+func New(options configapi.NodeConfig, server *kubeletoptions.KubeletServer) (*NodeConfig, error) {
 	if options.NodeName == "localhost" {
 		glog.Warningf(`Using "localhost" as node name will not resolve from all locations`)
 	}
@@ -87,15 +52,7 @@ func New(options configapi.NodeConfig, server *kubeletoptions.KubeletServer, pro
 		return nil, err
 	}
 
-	originClient, _, err := configapi.GetOpenShiftClient(options.MasterKubeConfig, options.MasterClientConnectionOverrides)
-	if err != nil {
-		return nil, err
-	}
-	kubeClient, privilegedKubeConfig, err := configapi.GetInternalKubeClient(options.MasterKubeConfig, options.MasterClientConnectionOverrides)
-	if err != nil {
-		return nil, err
-	}
-	externalKubeClient, _, err := configapi.GetExternalKubeClient(options.MasterKubeConfig, options.MasterClientConnectionOverrides)
+	externalKubeClient, kubeConfig, err := configapi.GetExternalKubeClient(options.MasterKubeConfig, options.MasterClientConnectionOverrides)
 	if err != nil {
 		return nil, err
 	}
@@ -104,25 +61,9 @@ func New(options configapi.NodeConfig, server *kubeletoptions.KubeletServer, pro
 	if err != nil {
 		return nil, err
 	}
-	clientgoClientSet, err := clientgoclientset.NewForConfig(privilegedKubeConfig)
+	kubeClient, err := clientgoclientset.NewForConfig(kubeConfig)
 	if err != nil {
 		return nil, err
-	}
-
-	if err = validateNetworkPluginName(originClient, options.NetworkConfig.NetworkPluginName); err != nil {
-		return nil, err
-	}
-
-	internalKubeInformers := kinternalinformers.NewSharedInformerFactory(kubeClient, proxyconfig.ConfigSyncPeriod.Duration)
-
-	// Initialize SDN before building kubelet config so it can modify option
-	var sdnNode network.NodeInterface
-	var sdnProxy network.ProxyInterface
-	if network.IsOpenShiftNetworkPlugin(options.NetworkConfig.NetworkPluginName) {
-		sdnNode, sdnProxy, err = NewSDNInterfaces(options, originClient, kubeClient, internalKubeInformers, proxyconfig)
-		if err != nil {
-			return nil, fmt.Errorf("SDN initialization failed: %v", err)
-		}
 	}
 
 	deps, err := kubeletapp.UnsecuredKubeletDeps(server)
@@ -142,7 +83,7 @@ func New(options configapi.NodeConfig, server *kubeletoptions.KubeletServer, pro
 	deps.KubeClient = externalKubeClient
 	deps.EventClient = kv1core.New(eventClient.CoreV1().RESTClient())
 
-	deps.Auth, err = kubeletapp.BuildAuth(types.NodeName(options.NodeName), clientgoClientSet, server.KubeletConfiguration)
+	deps.Auth, err = kubeletapp.BuildAuth(types.NodeName(options.NodeName), kubeClient, server.KubeletConfiguration)
 	if err != nil {
 		return nil, err
 	}
@@ -175,96 +116,13 @@ func New(options configapi.NodeConfig, server *kubeletoptions.KubeletServer, pro
 		AllowDisabledDocker: options.AllowDisabledDocker,
 		Containerized:       server.Containerized,
 
-		Client:                kubeClient,
-		ExternalKubeClientset: externalKubeClient,
-		InternalKubeInformers: internalKubeInformers,
-
 		VolumeDir: options.VolumeDirectory,
 
 		KubeletServer: server,
 		KubeletDeps:   deps,
-
-		ProxyConfig:    proxyconfig,
-		EnableUnidling: options.EnableUnidling,
-
-		SDNNode:  sdnNode,
-		SDNProxy: sdnProxy,
-	}
-
-	if enableDNS {
-		dnsConfig, err := dns.NewServerDefaults()
-		if err != nil {
-			return nil, fmt.Errorf("DNS configuration was not possible: %v", err)
-		}
-		if len(options.DNSBindAddress) > 0 {
-			dnsConfig.DnsAddr = options.DNSBindAddress
-		}
-		dnsConfig.Domain = server.ClusterDomain + "."
-		dnsConfig.Local = "openshift.default.svc." + dnsConfig.Domain
-
-		// identify override nameservers
-		var nameservers []string
-		for _, s := range options.DNSNameservers {
-			nameservers = append(nameservers, s)
-		}
-		if len(options.DNSRecursiveResolvConf) > 0 {
-			c, err := miekgdns.ClientConfigFromFile(options.DNSRecursiveResolvConf)
-			if err != nil {
-				return nil, fmt.Errorf("could not start DNS, unable to read config file: %v", err)
-			}
-			for _, s := range c.Servers {
-				nameservers = append(nameservers, net.JoinHostPort(s, c.Port))
-			}
-		}
-
-		if len(nameservers) > 0 {
-			dnsConfig.Nameservers = nameservers
-		}
-
-		services, err := dns.NewCachedServiceAccessor(internalKubeInformers.Core().InternalVersion().Services())
-		if err != nil {
-			return nil, fmt.Errorf("could not start DNS: failed to add ClusterIP index: %v", err)
-		}
-
-		endpoints, err := dns.NewCachedEndpointsAccessor(internalKubeInformers.Core().InternalVersion().Endpoints())
-		if err != nil {
-			return nil, fmt.Errorf("could not start DNS: failed to add HostnameIP index: %v", err)
-		}
-
-		// TODO: use kubeletConfig.ResolverConfig as an argument to etcd in the event the
-		//   user sets it, instead of passing it to the kubelet.
-		glog.Infof("DNS Bind to %s", options.DNSBindAddress)
-		config.DNSServer = dns.NewServer(
-			dnsConfig,
-			services,
-			endpoints,
-			"node",
-		)
 	}
 
 	return config, nil
-}
-
-func validateNetworkPluginName(originClient *osclient.Client, pluginName string) error {
-	if network.IsOpenShiftNetworkPlugin(pluginName) {
-		// Detect any plugin mismatches between node and master
-		clusterNetwork, err := originClient.ClusterNetwork().Get(networkapi.ClusterNetworkDefault, metav1.GetOptions{})
-		if kerrs.IsNotFound(err) {
-			return fmt.Errorf("master has not created a default cluster network, network plugin %q can not start", pluginName)
-		} else if err != nil {
-			return fmt.Errorf("cannot fetch %q cluster network: %v", networkapi.ClusterNetworkDefault, err)
-		}
-
-		if clusterNetwork.PluginName != strings.ToLower(pluginName) {
-			if len(clusterNetwork.PluginName) != 0 {
-				return fmt.Errorf("detected network plugin mismatch between OpenShift node(%q) and master(%q)", pluginName, clusterNetwork.PluginName)
-			} else {
-				// Do not return error in this case
-				glog.Warningf(`either there is network plugin mismatch between OpenShift node(%q) and master or OpenShift master is running an older version where we did not persist plugin name`, pluginName)
-			}
-		}
-	}
-	return nil
 }
 
 func buildCloudProvider(server *kubeletoptions.KubeletServer) (cloudprovider.Interface, error) {
